chore(base-driver): Group route definitions
