+++ packages/base-driver/lib/protocol/routes/appium/device.ts
@@ -1,0 +1,81 @@
+import type {Driver, MethodMap} from '@appium/types';
+
+/**
+ * Appium: Device interaction (apps, keyboard, files, clock).
+ */
+export const APPIUM_DEVICE_ROUTES = {
+  '/session/:sessionId/appium/device/system_time': {
+    GET: {command: 'getDeviceTime'},
+    POST: {command: 'getDeviceTime', payloadParams: {optional: ['format']}},
+  },
+  '/session/:sessionId/appium/device/activate_app': {
+    POST: {
+      command: 'activateApp',
+      payloadParams: {
+        required: [['appId'], ['bundleId']],
+        optional: ['options'],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/terminate_app': {
+    POST: {
+      command: 'terminateApp',
+      payloadParams: {
+        required: [['appId'], ['bundleId']],
+        optional: ['options'],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/app_state': {
+    POST: {
+      command: 'queryAppState',
+      payloadParams: {
+        required: [['appId'], ['bundleId']],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/install_app': {
+    POST: {
+      command: 'installApp',
+      payloadParams: {
+        required: ['appPath'],
+        optional: ['options'],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/remove_app': {
+    POST: {
+      command: 'removeApp',
+      payloadParams: {
+        required: [['appId'], ['bundleId']],
+        optional: ['options'],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/app_installed': {
+    POST: {
+      command: 'isAppInstalled',
+      payloadParams: {
+        required: [['appId'], ['bundleId']],
+      },
+    },
+  },
+  '/session/:sessionId/appium/device/hide_keyboard': {
+    POST: {
+      command: 'hideKeyboard',
+      payloadParams: {optional: ['strategy', 'key', 'keyCode', 'keyName']},
+    },
+  },
+  '/session/:sessionId/appium/device/is_keyboard_shown': {
+    GET: {command: 'isKeyboardShown'},
+  },
+  '/session/:sessionId/appium/device/push_file': {
+    POST: {command: 'pushFile', payloadParams: {required: ['path', 'data']}},
+  },
+  '/session/:sessionId/appium/device/pull_file': {
+    POST: {command: 'pullFile', payloadParams: {required: ['path']}},
+  },
+  '/session/:sessionId/appium/device/pull_folder': {
+    POST: {command: 'pullFolder', payloadParams: {required: ['path']}},
+  },
+} as const satisfies MethodMap<Driver>;
