+++ packages/base-driver/lib/protocol/routes/w3c/element.ts
@@ -1,0 +1,99 @@
+import type {Driver, MethodMap} from '@appium/types';
+
+/**
+ * W3C WebDriver: Elements (finding, state, and interaction).
+ * @see https://www.w3.org/TR/webdriver2/#elements
+ * @see https://www.w3.org/TR/webdriver2/#element-state
+ * @see https://www.w3.org/TR/webdriver2/#element-interaction
+ */
+export const W3C_ELEMENT_ROUTES = {
+  '/session/:sessionId/element/active': {
+    GET: {command: 'active'},
+  },
+  '/session/:sessionId/element/:elementId/shadow': {
+    GET: {command: 'elementShadowRoot'},
+  },
+  '/session/:sessionId/element': {
+    POST: {
+      command: 'findElement',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/elements': {
+    POST: {
+      command: 'findElements',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/element/:elementId/element': {
+    POST: {
+      command: 'findElementFromElement',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/element/:elementId/elements': {
+    POST: {
+      command: 'findElementsFromElement',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/shadow/:shadowId/element': {
+    POST: {
+      command: 'findElementFromShadowRoot',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/shadow/:shadowId/elements': {
+    POST: {
+      command: 'findElementsFromShadowRoot',
+      payloadParams: {required: ['using', 'value']},
+    },
+  },
+  '/session/:sessionId/element/:elementId/selected': {
+    GET: {command: 'elementSelected'},
+  },
+  '/session/:sessionId/element/:elementId/displayed': {
+    GET: {command: 'elementDisplayed'},
+  },
+  '/session/:sessionId/element/:elementId/attribute/:name': {
+    GET: {command: 'getAttribute'},
+  },
+  '/session/:sessionId/element/:elementId/property/:name': {
+    GET: {command: 'getProperty'},
+  },
+  '/session/:sessionId/element/:elementId/css/:propertyName': {
+    GET: {command: 'getCssProperty'},
+  },
+  '/session/:sessionId/element/:elementId/text': {
+    GET: {command: 'getText'},
+  },
+  '/session/:sessionId/element/:elementId/name': {
+    GET: {command: 'getName'},
+  },
+  '/session/:sessionId/element/:elementId/rect': {
+    GET: {command: 'getElementRect'},
+  },
+  '/session/:sessionId/element/:elementId/enabled': {
+    GET: {command: 'elementEnabled'},
+  },
+  '/session/:sessionId/element/:elementId/computedrole': {
+    GET: {command: 'getComputedRole'},
+  },
+  '/session/:sessionId/element/:elementId/computedlabel': {
+    GET: {command: 'getComputedLabel'},
+  },
+  '/session/:sessionId/element/:elementId/click': {
+    POST: {command: 'click'},
+  },
+  '/session/:sessionId/element/:elementId/clear': {
+    POST: {command: 'clear'},
+  },
+  '/session/:sessionId/element/:elementId/value': {
+    POST: {
+      command: 'setValue',
+      payloadParams: {
+        required: ['text'],
+      },
+    },
+  },
+} as const satisfies MethodMap<Driver>;
