+++ packages/base-driver/lib/protocol/routes/w3c/webauthn.ts
@@ -1,0 +1,46 @@
+import type {Driver, MethodMap} from '@appium/types';
+
+/**
+ * W3C Web Authentication (virtual authenticator automation).
+ * @see https://www.w3.org/TR/webauthn-2/#sctn-automation-add-virtual-authenticator
+ */
+export const W3C_WEBAUTHN_ROUTES = {
+  '/session/:sessionId/webauthn/authenticator': {
+    POST: {
+      command: 'addVirtualAuthenticator',
+      payloadParams: {
+        required: ['protocol', 'transport'],
+        optional: ['hasResidentKey', 'hasUserVerification', 'isUserConsenting', 'isUserVerified'],
+      },
+    },
+  },
+  '/session/:sessionId/webauthn/authenticator/:authenticatorId': {
+    DELETE: {
+      command: 'removeVirtualAuthenticator',
+    },
+  },
+  '/session/:sessionId/webauthn/authenticator/:authenticatorId/credential': {
+    POST: {
+      command: 'addAuthCredential',
+      payloadParams: {
+        required: ['credentialId', 'isResidentCredential', 'rpId', 'privateKey'],
+        optional: ['userHandle', 'signCount'],
+      },
+    },
+  },
+  '/session/:sessionId/webauthn/authenticator/:authenticatorId/credentials': {
+    GET: {command: 'getAuthCredential'},
+    DELETE: {command: 'removeAllAuthCredentials'},
+  },
+  '/session/:sessionId/webauthn/authenticator/:authenticatorId/credentials/:credentialId': {
+    DELETE: {command: 'removeAuthCredential'},
+  },
+  '/session/:sessionId/webauthn/authenticator/:authenticatorId/uv': {
+    POST: {
+      command: 'setUserAuthVerified',
+      payloadParams: {
+        required: ['isUserVerified'],
+      },
+    },
+  },
+} as const satisfies MethodMap<Driver>;
